Correctly handle scalar types, fix parse of numpy ints
Fixes: https://github.com/pytorch/pytorch/issues/29252

Summary:

There is some inconsistency in the code for parsing python numbers, e.g.
a function named THPUtils_checkDouble that returns true for integral
values as well as floating point values, and incorrect code in
python_arg_parser.

This led to issue #29252:

When we allow interpretation of a zero-dim numpy integer value
as a scalar in pytorch, we incorrectly parse the int as a float.

Test Plan:

Added a unit test based on the example given in the issue.

ghstack-source-id: 997dbffaf90f9ec2460d0bf30e55c85218fdaab3
Pull Request resolved: https://github.com/pytorch/pytorch/pull/30486