[BC-BREAKING] change index_select scalar_check to retain dimensionality of input.
The index_select documentaiton reads:
"The returned tensor has the same number of dimensions as the original tensor (input)."

But the implementation would return a 0-dimensional tensor iff both the input and index were 0-dimensional.
This change makes it so we retuan a 0-dimensional tensor iff the input is 0-dimensional.

ghstack-source-id: 1d7c1a82c63deeb0a2f11b96978838c74b8fa9e9
Pull Request resolved: https://github.com/pytorch/pytorch/pull/30502